feature(initial_pose): Option to set initial pose on challenge startup
I would like to combine some of the varying middleware components with
the challenge since they are challenge specific. By doing it this way,
we launch the challenge with one launch file together with the
dependencies. Setting the initial pose at challenge start is required
since we do not have time now to do this with rviz.